requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.0
colorama==0.4.6
selenium-wire==5.1.0
webdriver-manager==4.0.1
tqdm==4.66.1